LPS-179353 Extract ScreenNavigationEntry methods from CommerceTermEntryDetailsScreenNavigationCategory to CommerceTermEntryDetailsScreenNavigationEntry, extends CommerceTermEntryDetailsScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
